--- packages/prop-house-webapp/src/components/HouseManager/CreateTheRound/CreateTheRound.module.css
@@ -1,7 +1,0 @@
-.titleAndEditIcon {
-  align-items: flex-start;
-  justify-content: space-between;
-}
-.titleAndEditIcon svg {
-  margin-top: 12px;
-}
