--- packages/prop-house-webapp/src/components/HouseManager/RoundTiming/RoundTiming.module.css
@@ -1,5 +1,0 @@
-.addAddress {
-  display: flex;
-  gap: 0.75rem;
-  color: var(--brand-gray-semi-transparent);
-}
